preprocess/requirements.txt
numpy==1.24.4
pandas==2.2.1
scipy==1.13.0
biopython
h5py
